LPS-50050 Update usage of ReflectionTestUtil.invoke(Class, String, Class[], Object...)

+++ portal-impl/test/unit/com/liferay/portal/resiliency/spi/agent/HttpClientSPIAgentTest.java
@@ -559,7 +559,7 @@ public class HttpClientSPIAgentTest {
 		serializer.writeString(_SERVLET_CONTEXT_NAME);
 		serializer.writeObject(new SPIAgentRequest(_mockHttpServletRequest));
 
-		long receipt = (Long)ReflectionTestUtil.invoke(
+		long receipt = ReflectionTestUtil.invoke(
 			MailboxUtil.class, "depositMail", new Class<?>[] {ByteBuffer.class},
 			serializer.toByteBuffer());
 
@@ -945,7 +945,7 @@ public class HttpClientSPIAgentTest {
 			@Override
 			protected Datagram processDatagram(Datagram datagram) {
 				try {
-					long receipt = (Long)ReflectionTestUtil.invoke(
+					long receipt = ReflectionTestUtil.invoke(
 						MailboxUtil.class, "depositMail",
 						new Class<?>[] {ByteBuffer.class},
 						datagram.getDataByteBuffer());
@@ -978,7 +978,7 @@ public class HttpClientSPIAgentTest {
 		serializer.writeString(_SERVLET_CONTEXT_NAME);
 		serializer.writeObject(new SPIAgentResponse(_SERVLET_CONTEXT_NAME));
 
-		long receipt = (Long)ReflectionTestUtil.invoke(
+		long receipt = ReflectionTestUtil.invoke(
 			MailboxUtil.class, "depositMail", new Class<?>[] {ByteBuffer.class},
 			serializer.toByteBuffer());
